# Conflicts:
#	Enemy.cpp

+++ SnakeEnemyModel.h
@@ -32,7 +32,7 @@ public:
 private:
 	std::vector<TrailCollider*> colliderList;
 	std::vector<D3DXVECTOR3> moveTargetList;
-	std::vector<int> timeList;
+	std::vector<float> timeList;
 	std::map<TrailCollider*, std::list<std::weak_ptr<Enemy>>> colliderMap;
 
 	void SwapInColliderMap(TrailCollider* current, TrailCollider *next, std::shared_ptr<Enemy> enemy);
